chore: refactor code and combine data into single object
