+++ core/extensions/pdfFormatter.js
@@ -12,3 +12,24 @@ export const formatPropertyName = (str) => {
   const clearStr = str.replace(regex, '');
   return clearStr.split(' ').join('_').toLowerCase();
 }
+
+export const getColumnsData = (text, columnPropertyLine, columnDataStr) => {
+  const columnsStr = new RegExp(columnPropertyLine + ' 0 obj\\[<(.*?)>\\]', 's').exec(text)[1];
+  const columnPropertyRegex = /\/Name\((.*?)\)/g;
+  const columnDataRegex = /\((.*?)\)/g;
+  let columnPropertyMatch;
+  let dataMatch;
+  let dataArray = [];
+  let i = 0
+  let columnData = {};
+
+  while ((dataMatch = columnDataRegex.exec(columnDataStr)) !== null) {
+    dataArray.push(dataMatch[1]);
+  }
+
+  while ((columnPropertyMatch = columnPropertyRegex.exec(columnsStr)) !== null) {
+    columnData[formatPropertyName(columnPropertyMatch[1])] = getValueType(dataArray[i]);
+    i++;
+  }
+  return columnData;
+}
